io: move io_handle_read + TLS handshake helpers + record marker to handlers.c
Move into lib/io/handlers.c:
  - io_handle_read (the read completion handler)
  - process_record_marker (RPC record-marker reassembly)
  - handle_tls_handshake, process_ssl_accept, is_tls_client_hello,
    log_client_hello_details (static TLS helpers used by io_handle_read's
    ClientHello probe and handshake driver)
  - MAX_RPC_MESSAGE_SIZE define

All of these are backend-agnostic -- they operate on the buffer_state
and the conn_info's SSL pointer, never on io_uring or kqueue objects.
Remove the io_handle_read stub from backend_kqueue.c in the same
commit.

read.c shrinks to ~170 lines containing two liburing functions:
io_request_read_op (creates a fresh ic, submits read) and
io_resubmit_read (submits another read on an existing ic).
Everything else is now shared.

FreeBSD PR #7 note: the ClientHello probe in io_handle_read fires
on every fresh connection's first read.  If a probe matches but
reffs_server_ssl_ctx is NULL (as it is on FreeBSD PR #7 -- TLS not
initialized), handle_tls_handshake gracefully fails via SSL_new(NULL)
returning NULL and the caller closes the socket.  Not a crash; just
a misleading "Failed to create SSL" log for malformed clients.
Actual cleanup (short-circuit probe when ssl_ctx is NULL) deferred
to PR #8.

No semantic change on Linux: pure code move.
